+++ Tests/Integration/MSFT_xWinsSetting.config.ps1
@@ -1,0 +1,12 @@
+Configuration MSFT_xWinsSetting_Config {
+    Import-DscResource -ModuleName xNetworking
+
+    node localhost {
+        xWinsSetting Integration_Test
+        {
+            IsSingleInstance = 'Yes'
+            EnableLMHOSTS    = $Node.EnableLMHOSTS
+            EnableDNS        = $Node.EnableDNS
+        }
+    }
+}